Refactor(service): mover lógica de repository a service

+++ Backend/src/modules/categories/services/category.service.js
@@ -9,7 +9,11 @@ class CategoryService {
 
   async findAllCategories(page = 1, limit = 10) {
     try {
-      return await this.CategoryRepository.findAll(page, limit);
+      const categories = await this.CategoryRepository.findAll(page, limit);
+      if (!categories || categories.categories.length === 0) {
+        throw new Error("No se encontraron categorías");
+      }
+      return categories;
     } catch (error) {
       console.error("Error al obtener todas las categorías");
       throw error;
@@ -18,7 +22,11 @@ class CategoryService {
 
   async findCategoryById(id) {
     try {
-      return await this.CategoryRepository.findById(id);
+      const category = await this.CategoryRepository.findById(id);
+      if (!category) {
+        throw new Error("Categoría no encontrada");
+      }
+      return category;
     } catch (error) {
       console.error("Error al obtener la categoría");
       throw error;
@@ -27,7 +35,11 @@ class CategoryService {
 
   async findCategoryByName(nombre) {
     try {
-      return await this.CategoryRepository.findByName(nombre);
+      const category = await this.CategoryRepository.findByName(nombre);
+      if (!category) {
+        throw new Error("Categoría no encontrada");
+      }
+      return category;
     } catch (error) {
       console.error("Error al obtener la categoría");
       throw error;
@@ -38,7 +50,13 @@ class CategoryService {
     try {
       const { nombre, descripcion } = categoryData;
 
-      // Aplicamos el builder
+      // Validar si el nombre ya existe
+      const existingCategory = await this.CategoryRepository.findByName(nombre);
+      if (existingCategory) {
+        throw new Error("El nombre de la categoría ya está en uso");
+      }
+
+      // Builder para construir la categoría
       const builder = new CategoryBuilder()
         .setNombre(nombre)
         .setDescripcion(descripcion);
@@ -57,11 +75,32 @@ class CategoryService {
     try {
       const { nombre, descripcion } = categoryData;
 
+      // Verificar si la categoría existe
+      const existingCategory = await this.CategoryRepository.findById(id);
+      if (!existingCategory) {
+        throw new Error("Categoría no encontrada");
+      }
+
+      // Validar si el nombre ya existe
+      const existingCategoryByName = await this.CategoryRepository.findByName(
+        nombre
+      );
+      if (existingCategoryByName && existingCategoryByName.id !== id) {
+        throw new Error("El nombre de la categoría ya está en uso");
+      }
+
       const builder = new CategoryBuilder()
         .setNombre(nombre)
         .setDescripcion(descripcion);
 
-      const categoryToUpdate = builder.build();
+      let categoryToUpdate = builder.build();
+
+      // Filtrar undefined para no sobreescribir campos con null
+      categoryToUpdate = Object.fromEntries(
+        Object.entries(categoryToUpdate).filter(
+          ([_, value]) => value !== undefined
+        )
+      );
 
       return await this.CategoryRepository.update(id, categoryToUpdate);
     } catch {
@@ -72,6 +111,10 @@ class CategoryService {
 
   async deleteCategory(id) {
     try {
+      const category = await this.CategoryRepository.findById(id);
+      if (!category) {
+        throw new Error("Categoría no encontrada");
+      }
       return await this.CategoryRepository.delete(id);
     } catch (error) {
       console.error("Error al eliminar la categoría");
